requirements.txt
pydantic~=2.10.4
openai~=1.58.1
tenacity>=8.1.0,<9.0.0
pyyaml~=6.0.2
loguru~=0.7.3
numpy
datasets~=3.2.0
pydantic_core~=2.27.2
colorama~=0.4.6

html2text~=2024.2.26
pillow~=10.4.0
python-docx~=1.1.0
openpyxl~=3.1.2
beautifulsoup4~=4.12.3
pypdf~=4.1.0

# langchain==0.1.5
# langchain-text-splitters>=0.0.1,<0.1.0
# langchain-community==0.0.5
chromadb>=0.4.0,<0.7.0
pinecone-client~=3.1.0

gymnasium~=1.0.0
browsergym~=0.13.3
browser-use~=0.1.40
playwright~=1.49.1
googlesearch-python~=1.3.0

jsonpath-ng~=1.6.0
# jq~=1.6.0
pyarrow>=15.0.0
unidiff~=0.7.5
pandas~=2.2.0
opencv-python~=4.9.0.80
inquirerpy~=0.3.4

pydub~=0.25.1
python-magic~=0.4.27
python-socketio~=5.12.0
yt-dlp~=2023.12.30
youtube-search-python~=1.6.6
# tiktoken~=0.6.0

fastapi~=0.115.11
uvicorn~=0.34.0
flask>=2.0.1
aiofiles~=24.1.0
toml
tomli


pymysql>=1.1.0
cryptography>=41.0.4

cos-python-sdk-v5==1.9.25 


python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.25.0
jinja2>=3.1.2
websockets>=11.0.3

authlib>=1.2.1
itsdangerous>=2.1.2
